integracion_api_crm_servicios/requirement.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
pandas==2.2.3
paramiko==3.5.1
pip-chill==1.0.3
psycopg2-binary==2.9.11
pyodbc==5.3.0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
wheel==0.45.1
